Improve Makefile process to ensure that tests can run cleanly
Moved the PID file to a temporary directory that is automatically created
by the Makefile if needed (in an efficient manner).

Use free-port.pl to find a free TCP port to run the kssl_server on and use
that port for the test suite.

Wait for the kssl_server to become active (by checking for the PID_FILE)
before starting the test suite.