Fix libev/OpenSSL interaction in presence of WANT_READ/WANT_WRITE
In a recent refactor the handling of libev EV_READ and EV_WRITE events
and the possible OpenSSL error returns from SSL_read and SSL_write that
indicate that a retry is necessary (SSL_ERROR_WANT_READ and SSL_ERROR_WANT_WRITE)
were not being handled correctly.

Additionally, the code was assuming that an entire message (such as a payload)
could always be read in one go.