Complete port to libuv; resolve failure in test suite
The libuv documentation states that it's possible to get a UV_READABLE
when there's nothing to read. This appears to be happening causes us
to SSL_read when there's nothing to read and think there's an error.
Added a guard for this situation.

Also, reordered a uv_poll_stop() and a close() to ensure that there
is no polling of a closed file handle.