Switch from using polled file descriptors to using uv_stream
This is still working towards a threads rather than processes model so
that the port to Windows will be easier.

This commit remove the use of uv_poll against file descriptors and directly
access the TCP connection using uv_tcp functions. Data is read and written
to OpenSSL via in-memory OpenSSL BIO.

The current state is as follows:

1. The test suite passes.

2. --num-workers does nothing. All the fork() logic has been removed and
   everything is done in a single process. Next step will be to add threading
   and use --num-workers to specify the number of available threads.

3. Fails to detect that a connection has been closed and so the connection_state
   and all related memory structures are never freed: i.e. leaks memory like
   a sieve.