Document kssl_testclient options and add test-short Makefile target
The kssl_testclient parameters were previously undocumented and are
now documented in a comment at the start of kssl_testclient.c.

Added a test-short Makefile target that is the same as test but only
performs a small number of tests by setting the --short parameter on
the kssl_testclient. This can be handy for simple connectivity tests.