Address KEY-14. Implement versioning system used by RRDNS.
The makefiles Release and Version are a port of an number of shell scripts
used by RRDNS to create a versioning system. The key element is that each
time 'make release' is performed the minor version number is incremented
and the file RELEASE_NOTES is updated with all changes to keyless since
the last time a release was performed. This achieved by reading a tag
from git.