Pad good/bad response to 4,096 bytes
A new tag type, KSSL_TAG_PADDING, and is used to pad operations (via
the flatten_operation function) and error messages (via the kssl_error
function) to the same length (4096 + KSSL_ITEM_HEADER_SIZE) bytes.

This is done so that an eavesdropper cannot use the length of a
returned item as a way to detect an error.