Fix broken pipelining when more than 16 messages queued
The pending outbound message queue has a maximum of 16 entries after
which data is lost. It is never meant to fill up. The pipelining
change meant that queued messages were not flushed until all the
incoming requests had been processed. Sending more than 16 requests
on a single connection would break things. This showed up as the
pipeline tests not passing.